Add functions for managing JSON.
These make it easier to write code to generate JSON. While at it, use them to
generate the JSON for `mirrors.json`. To make it so that `array_to_string()`
omits columns with `NULL`s in them (rsync and now notes), set up a fallback
value for the `json_value()` functions. When it's `NULL`, that's what gets
returned, and therefore `array_to_string()` ignores it.

I may or may not also modify the JSON-handling code for distributions to use
thse functions. It would be nice to have the keys in the same order every
time.