Add `by_extension_json()`.
The idea is that the app will call it after a distribution is added in order
to update all the `/by/extension/*` JSON files for each extension in the
distribution. The function returns a row for each extension with two columns:
The extension name and the JSON to be output to the file.

I'm wondering now if I'm being too strict in requiring that every extension
have a unique version number in every distribution. CPAN doesn't require this;
one can upload multiple distributions with extensions with the same version
numbers, just as long as the distribution version number is unique. Think
I'll leave it for now, seems like a good idea anyway.