Add middleware options and proper db username to config.
This is in the JSON configuration files. All of them should connect with the
"pgxn" user. To build the database, we use the `--db_super_user` option to
`Build.PL` or `./Build db`.

The configurable middleware is also in these files. Test loads `StackTrace`,
staging and prod load `ErrorDocument`, and dev loads `StackTrace`, `Debug`,
and `Lint`.