Test router authentication.
Done by adding a class for testing, TxnTest. This class sets up the
environment so that everything is in a single transaction rolled back at the
end. The transaction methods of the DBI are mocked out. It also has
convenience functions for creating an admin user and a non-admin user.