Handle archive error conditions.
Bit of a hack to trap Archive::Extract errors; Archive::Zip is a bit better.
Set up the error handlers for each to throw an array for an exception, where
the only item in the array is a `sprintf` format that we can then use to fill
in the file name. This will be useful for sending an error back to the user on
upload.