New JSON format for extensions.
This is to deal with the same version of an extension appearing in multiple
distributions. I can't tell you what a PITA this was. The function,
`by_extension_json()`, is a bit of a monstrosity, although it at least
executes only one query. I blame Aristotle.

Anyway, I think it's a good format. Now to fix the bug with the tag JSON.