Add Ajax sugar to moderate screen.
The Accept/Reject links now do async requests. If you have JS disabled, they
still work by the old-fashioned GET request and reload.

To make error-handling work nicely in the browser, I've added a new method to
the controller, `respond_with()`. Pass it the name of an error status and if
it's an XHR request, it responds appropriately. So the above Ajax requests ask
for HTML, and if there's an error, the pop the HTML error mesage on the
screen. Need to write some tests for it.

Some other changes:

* Check the return value of the call to `set_user_status()` and return a 404
  if it returns false. The assumption is that if it doesn't update a record,
  it couldn't find one to update.

* Rename the "403" template to "forbidden". Need to add tempaltes for
  "notfound" and "success" too.

* Comment out the body of the `upload` action. Might as well just ignore it
  till I write it.