Test localization distribution conflict exception.
Required a bit of tweaking of the mesage in the database. Also had to tweak
TxnTest to reset the `$user` and `$admin` variables in `reset` so that they
will be recreated as appropriate.

While at it, I removed a bunch of exception handling for extension version
conflicts, as that constraint was removed in
8ecee519615043573554f4d975d4db9f73abbce5.