Redirect with 303 on successful POST.
I hadn't noticed this HTTP 1.1 error code until confound pointed it out to me
on #plack. It's specifically for use redirecting after a POST.

I was going to have a failed POST also redirect, but in discussion, it seems
to make more sense not to redirect, as the reason for the redirect is to
prevent a double-submission, but we actually *want* the user to correct the
form and submit again. So I think 409 is appropriate there.

While at it, remove unneeded `$code` variable, as it was always set to 409
anyway.