Mount /auth and /pub as separate apps.
The changes mostly here mostly involve URI rejiggering. All URIs now need to
be generated without a leading /auth or /pub passed to `uri_for()`. Also had
to mount `/ui` under both apps, and add routes for /about and /contact to the
/auth app, too.

I think the only thing not working properly is links between the two sites.
Really just the link from the public site to login to the private site. Need
to think about how to do that.