Add `uri_script_name_key` configuration directive.
This is used by `Request->uri_for()` to determine the script name. The reason
for this is so that a proxy server can set a header for a different script
name that represents the proxy's root, rather than the app's root. It has to
be enabled by someone who is setting the header on the server, of course.