Add `auth_uri_for()`.
Should only be called from the public site, and all URIs are treated as
relative to `auth_uri`. Required some reworking of `auth_uri` when not from
the config file. Seems good now. Will apply it in some more places tomorrow,
and move the password reset stuff to the auth URI, too.