Improve error handling.
Users must configure these middleware:

        ["ErrorDocument", 500, "/error", "subrequest", 1],
        ["HTTPExceptions"],
        ["StackTrace", "no_print_errors", 1]

StackTrace generates the stacktrace. Hopefully my [pull
request](http://github.com/miyagawa/Plack/pull/144) will be accepted so that
the other middleware can easily access it.

HTTPExceptions catches the exception and turns it into a 500 response.

ErrorDocument then sends a subrequest to /error. The controller for this URI
sends an email with information about the failed request, including a copy of
the environment and the stack trace (depending on that pull request, of
course). Then it sends a friendly 500 error page to the user.