Add "Add" and "Delete" buttons to mirror list.
Need to link to a page to create a new mirror, so do that. Also add a "Delete"
button to the action column. Will be handy, and will use DELETE Method. :-)

While at it, improve the table layout CSS. I think the text alignment is a lot
better now, at least in Safari.