Add interface to add a new mirror.
Doing basic validation in JavaScript, before inserting into the database, and
upon error inserting into the database. Missing or invalid fields are
highlighted (or listed for an XHR request). A couple of missing fields were
added to the new mirror interface, too: Organization and Rsync. And the width
of form labels and fields has been made consistent across the app.

Tests next.