Resolve contradicting meanings of "owner".
The term "owner" is now used only in the context of permissions: who owns or
co-owns.a particular extension. The user who uploads a distribution is now
recorded as the "creator" of that extension, and listed in the JSON metadata
simply as the "user". The term "creator" is not used here because the user who
created the distribution (that is, uploaded it) should not be confused the the
person who created the code in the distribution.

To enact this change in terminology, this commit makes these changes to the
code and its outputs:

* Rename the `distributions.owner` column to `distributions.creator`.
* Rename the "owner" key in the distribution metadata file to "user".
* Rename the "owner" attribute in PGXN::Manager::Distribution to "creator".
* Move the `/by/owner/` metadata files to `/by/user/`.
* Rename the `by_owner_json()` SQL function to `by_user_json()`.
* Change the name of the URI template for the user JSON output by
  `add_distribution()` to `by-user`.