Fix a few extraction-related bugs.
First, require HTTP::Body 1.08. That's the module that creates the temp files
used for uploads, and 1.08 properly appends the file name suffix to the temp
file, which allows Archive::Extract to figure out what the archive type is.
With older versions of HTTP::Body, if a non-zip file was uploaded,
Archive::Extract could not determine what the archive format was, and died.

Relatedly, properly catch such a death. One of the downsides of Try::Tiny that
I keep forgetting is that it uses subroutine references. So a `return` returns
from those references, *not* from the method that's calling them. That's a
significant difference from the behavior of `eval` or, indeed, of a
hypothetical `try` keyword in Perl itself. Anyway, the fix is to always return
from the blocks passed to `try`, and to use the return value of `try` as the
return value of the method. This is what allowed me to see that
Archive::Extract was getting confused, whereas before it just ate the error.

As a benefit, `extract()` now properly returns false on failure. The tests
were wrong before, expecting it to return true. Guh.

Closes #4.