Add "doc" key to "provides" map.
In passing, fix a few typos found in re-reading the whole doc and removed a
few gratuitous mentions of PostgreSQL. It's surprisingly stable document; the
addition of the "doc" key is the only thing that needs to change since I
developed master, api, and site. The only other thing I might want to change
is to remove the PostgreSQL-specific stuff altogether, since PGXN need not be
PostgreSQL-specific.