Lowercase dist, extension, and version in file names.
Distribution and extension names, as well as version strings, should now
always be converted to lowercase before writing files to disk. This is because
all these values (as well as tags and nicknames, which were already lowercased
in file names) are case-insensitive. So while preserving case in JSON files,
force lowercase in file names for consistency.