Fix encoding issues.
Either DBD::Pg 3.0 or Perl 5.20 is stricter about things, but encodings were
messed up in a couple of places. So use `encode()` and `decode()` instead of
`encode_utf8()` and `decode_utf8()` for stricter encoding enforcement, and add
a couple of decodings that were missing: username and password for basic auth,
and reading a file from Archive::Zip.

Also fix a couple of warnings for `returns` that were never reached, and
from Twitter::Lite requiring SSL.