Allow archive file args to `pgxn_maint reindex`.
Add the ability to pass the path to a zip archive to `pgxn_maint index`.
Previously, you had to pass a distribution name and release version (and you
still can), but that required that one manually place the archive where the
existing one is, first. This is cleaner, as now `pgxn_maint` will process the
file from wherever you have it and put it where it belongs.