Check versions on upload
When adding a new release to the database, make sure the distribution version
is greater than any previous version, and that all extension versions are
greater than or equal to the extension as seen before. Required rejiggering
the tests which had been adding a lower-versioned extension than before! But
still need to add this functionality to `update_distribution()` and to write
tests.