Test and fix version check functions.
Need to be more comprehensive checking extension versions. It doesn't matter
what distribution they're in, only that we check for a version lower than in
previous releases or higher than in later releases. So abandon looking at a
specific distribution and version, and instead rely entirely on the date. In
`update_distribution()`, we pass the date of the release by looking it up in
the `distributions` table. In `add_distribution()`, we only need to look at
all previous releases, so no need to look at the date.

So now there are two extension version-checking functions:
`check_prev_versions()` and `check_later_versions()`. The former is called by
both `add_distribution()` and `update_distribution()`, while the latter is
called only by `update_distribution()`. And there are tests for these
new functions!