More consistently recognize upload types
Sometimes Plack writes an upload file with a completely randomized name,
without the original file basename. Because Archive::Extract uses the archive
file name to detect the type of archive, it fails on this temporary name. I've
submitted jib/archive-extract#8 to refactor type detection into a separate
function to call, but in the meantime, I've duped it here and added some tests
to ensure it works as expected. Resolves #61.