Remove extension version checks
With the ability to create new releases of lower versions added in d2bd3bf,
the extension version checks became broken, as they were always ensuring that
versions were higher than before. I could implement a bunch of rules to check
for this and that, but they would have to be pretty loose --- and more likely
to annoy than help. They would also be cumbersome, difficult to maintain, and
complicated. So remove all the extension version checks, and just rely on
authors to version things correctly. The distrbituion version checks are
probably sufficient to prevent most problems.