Simplify and correct uri_for()
It was designed to handle paths relative to the current path, but
auth_uri_for() never did, and when some of the instances of auth_uri_for()
were converted to uri_for() in 31e5145, they sometimes passed a full path
without a leading slash, leading to invalid paths. However, thre were
almost no uses of uri_for() without a full path.

So remove support for relative paths in uri_for() and ensure that full paths
are always passed to it. Furthermore, simplify it to properly handle the
script path prefix, whether forwarded by a proxy or not; previously it was a
bit confused about the difference between a script path (the prefix for the
app itself) and a request path. And better document and properly test these
variants.