Fix processing zip files with no prefix.
I noticed that PostgreSQL_Faker was borked; it contained a file named
`postgresql_faker-0.4.0` but no `META.json`. Damien Clochard send me the file
he originally uploaded, which I have added here in a test case, and I was able
to replicate and correct the issue. Just need to be consistent in the handling
of prefixes: with no prefix, there is no slash before file names! So ignore
the slash when looking for a prefix.