Suppress trace and args in the trace
Just put the error message in the admin email on error, not the full trace, and
remove arguments from the trace, too. This reduces noise in the email and prevents
leaking secrets like passwords. It's a bit hinky to pull the erorr message out of the
HTML <title> element generated by the StackTrace middleware but ¯\_(ツ)_/¯.