Fix Postgres 16 issues
Tests failed because sometimes a tag appears in tag stats as `Foo` and
other times as `foo`. So teach the tag stats to always lowercase tag
names for consistent formatting, both in tests and as things change in
time.

Replace the PL/Perl-based mocks, which used the `%_SHARED` hash, with a
session-based config value. Required because security appears to have
been increased on `%_SHARED`, as it was empty inside the `SECURITY
DEFINER` functions that called mock functions that use it. Since we
require Postgres 12 or higher now, we can do away with that hack ans use
the much cleaner configuration-based mocks.

Update the CI workflow to test on Postgres 16 and use the most recent
checkout action.

While at it, eliminate stray diagnostic lines from test output from the
pgTAP tests by using a different unprintable character in the
`distributions.pg` test, and using error descriptions instead of the
invalid terms and tags in `types.pg`.