Move PID deletion out of DEMOLISH method
There is no guarantee from Perl that it's called in a timely manner or
order when shutting down. The `pid_file` field may no longer be set.
This might explain why the PID file was never cleaned up.

Move it instead to the `run` method, where it should always run once a
signal handler has set `continue` to false. It exits the loop on the
next run, then cleans up the PID file before exiting. At least that's
the idea.

Move the PID test to the `run` test section of `t/consumer.t`, since
`run` is where it happens now.