Try to shut down more aggressively
Add a separate `_shutdown` method called by `run` and by signal
handlers. It's responsible for removing the PID file and for setting
`continue` to `0`. This should ensure that it's always called at least
once, and is safe to call multiple times. Either way the PID file should
now always be deleted and the shutdown activity properly logged.

Also set all the signal handlers in the `_signal_handlers` method, and
teach it to nest signal handlers if necessary. This will keep the
consumer from wiping out its own handlers. I have no reason to think
they have been, but it seems best to play it as safe as possible.

Add a slew of new tests to test all these behaviors thoroughly.