Move howto to Locale and script to format HTML
Rather than manually regenerate it and store it in %Lexicon, ship it as
an `.html` file in the appropriate locale directory. Pilfer the
`from_file` Locale method from PGXN::Site, but remove the
Locale::Maketext processing step to avoid the need to escape characters
it looks for. There should be no Locale::Maketext processing anyway:
just the translated document. Fixes #76.

Add `bin/format_l10n_docs`, a simple script that uses `multimarkdown` to
regenerate the `.html` version of Markdown files in `lib`. This will
allow other localize-able documents to be added in the future.