Fix warning, rsync URL validation
Remove the quotation marks from ` HTTP::Body '1.08'` to fix a warning
new in Perl 5.40. I apparently have a lot of code that incorrectly
quotes versions.

Add a custom validation function to the jQuery Validation plugin to
validate rsync URLs. Apparently at some point the url validator included
in jQuery Validation was limited to checking http and ftp URLs. This
prevents the mirror form from failing to submit because of an incorrect
validation error. So crib from the [url validation] to create an rsync
URL validator.

Discovered after fixing the Content Security Policy header on
manager.pgxn.org. Apparently inline JS hasn't worked since introducing
the header some time ago, because JS code in `<script>` elements was
prevented from executing. Fixing the header is what broke the mirror
form.

So add comments to the README on how to properly set the CSP header, and
how to generate the appropriate SHA-256 digests for it.

Add a hint to the README for building Compress::Raw::Lzma, which I've
had to figure out several times now. Hopefully this will save a little
time in the future.

  [url validation]: https://github.com/jquery-validation/jquery-validation/blob/master/src/additional/url2.js