Add support for PostgreSQL 17-18
PostgreSQL 17 added json_value() and made "json" a reserved word. So
update the relevant SQL scripts and tests to rename `json_value()` to
`into_json()` and to quote `"json"` when used as a function parameter or
return table column name. Add `sql/20-into-sql` to simply re-execute
those files, so that an existing database will be updated ahead of
upgrading to PostgreSQL 17 or later.

Also:

*   Add PostgreSQL 17 and 18 to the CI test workflow
*   Explicitly load `POSIX` in `t/mastodon.t`
*   Increment to v0.33.0 and timestamp for release