Add `21-cleanup-json.sql` and pg 17 upgrade notes
Add a new migration, `sql/21-cleanup-json.sql`, that drops the old
`json_value` function that was renamed in `sql/20-into-json.sql`.

Add the "Upgrading to PostgreSQL 17" section to the README to note that
the database must be dumped with the `--quote-all-identifiers` option to
`pg_dump` in order to compatibly migrate to PostgreSQL 17.

Increment to and timestamp v0.33.1.